feat: smooth float-based movement with delta time
- Entity/Player/Monster: int coords → float
- IRenderer/Renderer.h: drawChar signature → (float, float)
- Engine: added deltaTime + playerSpeed for frame-independent movement
- WASD/Arrows for movement, Q to quit
- Map: isWalkable/spillBlood accept float

+++ include/rogue/Renderer.h
@@ -8,6 +8,6 @@ public:
   virtual void shutdown() = 0;
   virtual void clear() = 0;
   virtual void refresh() = 0;
-  virtual void drawChar(int x, int y, char ch, int color) = 0;
+  virtual void drawChar(float x, float y, char ch, int color) = 0;
 };
 } // namespace rogue
