+++ include/rogue/entities/Monster.h
@@ -5,7 +5,7 @@ namespace rogue {
 
 class Monster : public Entity {
 public:
-  Monster(int x, int y, char symbol, int color, int maxHp);
+  Monster(float x, float y, char symbol, int color, int maxHp);
   void update() override; // Implement monster-specific behavior
 };
