+++ include/rogue/entities/Player.cpp
@@ -2,14 +2,14 @@
 
 namespace rogue {
 
-Player::Player(int x, int y, char symbol, int color, int maxHp)
+Player::Player(float x, float y, char symbol, int color, int maxHp)
     : Entity(x, y, symbol, color, maxHp) {}
 
 void Player::update() {
   // Implement player-specific behavior (e.g., handle input)
 }
 
-void Player::move(int dx, int dy) {
+void Player::move(float dx, float dy) {
   x += dx;
   y += dy;
 }
